package/petitboot: Change the boot.d/ hooks
Replace the hooks in boot.d/. Do not install any hooks that were
distributed with petitboot and instead add a single hook into
the fs-overlay. This remaining hook does a copy of the dtb, even if that
dtb is user-provided. This wasn't the case with the create-default-dtb
hook from petitboot, and is important if later hooks want to modify
the dtb (as is the case with petitboot-m1-hook).

Also source the default dtb from /sys/firmware/fdt instead of
/proc/device-tree, as the latter loses some information.